Move sync topics command to taxonomy app
